adding inference file
